This is the start of repeated field support.
Very messy work-in-progress of adding array support to cpp-parquet.
I've modified the AvroSchemaWalker to detect and support ARRAY
types (only of INTS for now) in JSON schemas, checked in an example JSON
schema with an ARRAY of INT fields, and copied some code from Impala
that will do the actual RLE & bitpacking of the repetition levels.
However, none of it builds or writes any repetition levels right now.