Large data support/test and some fixes
I’ve added support for a column with 2GB (gibibytes) of data as well as
a test case.  However this exposed a bug in that we don’t support
multiple pages (in fact I was going to defer on this as long as
possible because it’s not clear how to calculate the data size
requirements without iterating through each piece of data in the
column, looking at the R&D levels, and also potentially compressing the
data and, if it’s too large, uncompressing it and picking the “right”
cut off, which seems hard to do on uncompressed data).  But it looks
like multiple pages is a must-have because the JAVA readers allocate a
byte array of the length of the page, and JAVA arrays cannot be larger
than INT_MAX elements in length.
Sigh.