Add record metadata tracking
This is just the beginning, but the idea is that we need to track the
indices into the R&D level array, as well as the data buffer for each
record added to the column.  Eventually this data will be used to
figure out how many & which subset of this column’s data are flushed in
each row group.

Also removed the {begin,end}_index code I’d added, as it wasn’t a great
approach (and assumed that the R&D vector indices as well as the data
buffer indices were the same, and they clearly are not)