More complete schema walking code
I’ve restructured all of it:

-Remove union_from_null, and push all parquet-specific transformations
into the callback, rather than the schema walker.  For instance, the
“compression” of a union subtree (3 nodes) into one ParquetColumn
object is handled inside AtNode().  To do this I had to add setters for
the repetition type and data type to ParquetColumn.  The reason is for
this is that I didn’t want separate logic for constructing
ParquetColumn objects for each of “union { null, int }” and “int”
fields.  This way, the bottom of the Avro schema tree (the int) field
is set to OPTIONAL as the recursion down the schema is undone. The same
will happen for ARRAY fields.
-Added an std::map for the type mapping between AVRO and Parquet.
-Added a contract for the schema walking so that the callback can
return false if further recursion shouldn’t be done.  This is used
inside the UNION handling because the children of a UNION node are
handled specially (the returned node from the child of a UNION is set
to OPTIONAL)
-Added a BUNCH of code that needs to be removed (marked with TODO) that
are mostly around getting BOOLEANS to get through the schema
conversion, so I can test it with a real schema from ADAM.